fixed scene loading issues
fixed issue with uploading empty scenes
fixed issue with dynamic objects being written incorrectly when changing
scenes
fixed issue with dynamic snapshot properties not being sent correctly
removed unused preferences

+++ CognitiveVRUnity/Assets/CognitiveVR/Scripts/SensorRecorder.cs
@@ -50,7 +50,7 @@ namespace CognitiveVR
             sb.Append(",");
             JsonUtil.SetString("sessionid", CognitiveVR_Preferences.SessionID, sb);
             sb.Append(",");
-            JsonUtil.SetDouble("timestamp", CognitiveVR_Preferences.TimeStamp, sb);
+            JsonUtil.SetDouble("timestamp", (int)CognitiveVR_Preferences.TimeStamp, sb);
             sb.Append(",");
             JsonUtil.SetInt("part", jsonPart, sb);
             sb.Append(",");
